BACKLOG-16051 Remove previous table, rename a few files, clear logs, add local storage
