--- src/javascript/JContent/ContentRoute/ContentLayout/ContentListTable/ContentListEmptyDropZone.jsx
@@ -1,43 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import {TableBody, TableRow, TableCell, withStyles} from '@material-ui/core';
-import EmptyDropZone from '../EmptyDropZone';
-import UploadTransformComponent from '../UploadTransformComponent';
-
-const styles = theme => ({
-    dragZoneContentList: {
-        position: 'absolute',
-        height: '57vh',
-        width: '100%',
-        display: 'flex',
-        padding: theme.spacing.unit * 4
-    },
-    dragZone: {
-        display: 'flex',
-        padding: 0,
-        width: '100%'
-    }
-});
-
-export const ContentListEmptyDropZone = ({classes, path, mode}) => (
-    <TableBody>
-        <UploadTransformComponent
-            uploadTargetComponent={TableRow}
-            uploadPath={path}
-            className={classes.dragZoneContentList}
-            mode={mode}
-        >
-            <TableCell className={classes.dragZone}>
-                <EmptyDropZone component="div" mode={mode}/>
-            </TableCell>
-        </UploadTransformComponent>
-    </TableBody>
-);
-
-ContentListEmptyDropZone.propTypes = {
-    classes: PropTypes.object.isRequired,
-    mode: PropTypes.string.isRequired,
-    path: PropTypes.string.isRequired
-};
-
-export default withStyles(styles)(ContentListEmptyDropZone);
