--- src/javascript/JContent/ContentRoute/ContentLayout/ContentListTable/ContentListTable.constants.js
@@ -1,5 +1,0 @@
-const ContentListTableConstants = {
-    appTableCells: 2
-};
-
-export default ContentListTableConstants;
