--- src/javascript/JContent/ContentRoute/ContentLayout/ContentListTable/index.js
@@ -1,3 +1,0 @@
-import ContentListTable from './ContentListTable';
-
-export default ContentListTable;
