--- src/javascript/JContent/ContentRoute/ContentLayout/ContentListTable/todo.md
@@ -1,5 +1,0 @@
-* remove table switcher (to switch use {type: 'tableType', payload: 'moon'})
-* rename *Moon components (remove Moon)
-* restore package json linting
-* remove unnecessary lint disable
-* remove this file
