+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentTable/ContentTable.constants.js
@@ -1,0 +1,5 @@
+const ContentTableConstants = {
+    appTableCells: 2
+};
+
+export default ContentTableConstants;
