+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentTable/ContentTypeSelector/ContentTypeSelector.scss
