+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentTable/EmptyTable.jsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {Table, TableBody, TableBodyCell, TableRow, Typography} from '@jahia/moonstone';
-import ContentListTableConstants from './ContentListTable.constants';
-import ContentListTableWrapper from './ContentListTableWrapper';
+import ContentTableConstants from './ContentTable.constants';
+import ContentTableWrapper from './ContentTableWrapper';
 
 const EmptyTable = ({columnSpan, t}) => (
-    <ContentListTableWrapper>
+    <ContentTableWrapper>
         <Table aria-labelledby="tableTitle" data-cm-role="table-content-list">
             <TableBody>
                 <TableRow>
-                    <TableBodyCell colSpan={columnSpan + ContentListTableConstants.appTableCells + 2}>
+                    <TableBodyCell colSpan={columnSpan + ContentTableConstants.appTableCells + 2}>
                         <Typography weight="bold">{t('jcontent:label.contentManager.noResults')}</Typography>
                     </TableBodyCell>
                 </TableRow>
             </TableBody>
         </Table>
-    </ContentListTableWrapper>
+    </ContentTableWrapper>
 );
 
 EmptyTable.propTypes = {
