+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentTable/index.js
@@ -1,0 +1,3 @@
+import ContentTable from './ContentTable';
+
+export default ContentTable;
