+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentTable/reactTable/components/cells/CellStatus.jsx
