+++ src/javascript/JContent/ContentRoute/ContentLayout/useKeyboardNavigation.js
@@ -18,6 +18,7 @@ export const useKeyboardNavigation = ({listLength, onSelectionChange}) => {
     };
 
     const setFocusOnMainContainer = () => {
+        console.log(mainPanelRef);
         mainPanelRef.current.focus();
     };
